seperate modules for each authentication method
